+++ dnx_webui/source/main/dfe_main.py
@@ -84,8 +84,6 @@ from source.system.dfe_services import WebPage as dnx_services
 
 from source.main.dfe_authentication import Authentication, user_restrict, update_session_tracker, authenticated_session
 
-import source.messenger.msg_main as messenger
-
 # ===============
 # TYPING IMPORTS
 # ===============
@@ -677,7 +675,9 @@ def handle_system_action(page_settings: dict):
 
     return render_template('main/device.html', theme=context_global.theme, **page_settings)
 
+# =================================
 # HELPERS
+# =================================
 def get_default_page_settings(session_info, *, uri_path: list[str]) -> dict:
     '''sets the following values:
 
@@ -716,7 +716,7 @@ def user_timeout() -> None:
     session.modified = True
 
 @app.before_request
-def set_user_settings(user: str) -> None:
+def set_user_settings() -> None:
     if not (user := authenticated_session()): return
 
     # ------------------
@@ -727,7 +727,7 @@ def set_user_settings(user: str) -> None:
         if new_theme not in ['light', 'dark']: return
 
         with ConfigurationManager('logins', file_path='/dnx_webui/data') as webui:
-            webui_settings = webui.load_configuration()
+            webui_settings: ConfigChain = webui.load_configuration()
 
             # this check prevents issues with log in/out transitions
             if user in webui_settings.get_list('users'):
@@ -737,7 +737,7 @@ def set_user_settings(user: str) -> None:
                 webui.write_configuration(webui_settings.expanded_user_data)
 
 @app.before_request
-def load_user_settings(user: str) -> None:
+def load_user_settings() -> None:
 
     # loading defaults before returning. manually setting since theme is only tracked setting as of now.
     if not (user := authenticated_session()):
@@ -750,163 +750,21 @@ def load_user_settings(user: str) -> None:
         context_global.settings = web_config.get_dict(f'users->{user}->settings')
 
 
+# =================================
+# WEBUI EXTENSIONS
+# =================================
+# webui template registrations
+import source.main.dfe_template_globals
 
-# ================
-# THEMES
-# ================
-theme_common = {
-    'mode': '',
-    'nav_text': 'blue-grey-text text-darken-2',
-    'subnav_text': 'blue-grey-text text-darken-3',
-    'tab_text': 'blue-grey-text text-lighten-2',
-    'tab_classes': 'tab col s4 l3 xl2',
-    'icon': 'teal-text text-lighten-2',
-    'modal_text': 'blue-grey-text center',
-}
+# webui themes
+import source.main.dfe_themes
 
-theme_dark = {
-    'background': (
-        'style="background: url(static/assets/images/dnx_bg1_dark.svg); '
-        'background-repeat: repeat; '
-        'background-attachment: fixed;"'
-    ),
-    'main_section': 'blue-grey lighten-2',
-    'off_section': 'blue-grey lighten-5',
-    'card': 'blue-grey lighten-4',
-    'title': 'black-text'
-}
+# secure messenger extension
+import source.messenger.msg_main
 
-theme_light = {
-    'background': (
-        'style="background: url(static/assets/images/dnx_bg1_light.svg); '
-        'background-repeat: repeat; '
-        'background-attachment: fixed;"'
-    ),
-    'main_section': 'grey lighten-2',
-    'off_section': 'grey lighten-5',
-    'card': 'grey lighten-4',
-    'title': 'blue-grey-text text-darken-1'
-}
-@app.before_request
-def set_theme_values() -> None:
-    style = context_global.settings['theme']
-
-    context_global.theme = {'mode': style}
-    context_global.theme.update(theme_common)
-
-    if (style == 'dark'):
-        context_global.theme.update(theme_dark)
-
-    elif (style == 'light'):
-        context_global.theme.update(theme_light)
-
-# ====================================
-# FLASK API - TEMPLATE FUNCTIONS
-# ====================================
-@app.template_global()
-def create_title(title: str) -> str:
-    return (
-        f'<div class="row"><h4 class="{context_global.theme["title"]} card-title">{title.title()}</h4></div>'
-        f'<div class="title-divider"></div><br>'
-    )
-
-@app.template_global()
-def create_switch(label: str, name: str, *, tab: int = 1, checked: int = 0, enabled: int = 1) -> str:
-    if (not enabled): status = 'disabled'
-    elif (checked): status = 'checked'
-    else: status = ''
-
-    return ''.join([
-        f'<form method="POST"><input type="hidden" name="tab" value="{tab}">',
-        f'<div class="input-field col s6 center">{label}<div class="switch"><label>Off',
-        f'<input type="checkbox" class="iswitch" name="{name}" {status}>',
-        '<span class="lever"></span>On</label></div></div></form>'
-    ])
-
-@app.template_global()
-def create_tab(active_tab: int, cur_tab: int, href: str) -> str:
-    tab = (
-        f'<li class="{context_global.theme["tab_classes"]}">'
-        f'<a href="#{href}" onclick="activeTab({cur_tab})" class="{context_global.theme["tab_text"]}'
-    )
-
-    if (cur_tab == active_tab):
-        tab += ' active'
-
-    tab += f'">{href.replace("-", " ").title()}</a></li>'
-
-    return tab
-
-@app.template_global()
-def create_button_with_modal(
-        classes: str, icon: str, index: int, num: int, tab: int, btn_name: str, btn_value: str, message: str) -> str:
-
-    btn_classes = f'{classes} waves-effect waves-light modal-trigger'
-
-    button = (
-        f'<a class="{btn_classes}" href="#modal{index}-{num}"><i class="material-icons">{icon}</i></a>'
-        f'<div id="modal{index}-{num}" class="modal">'
-          f'<div class="modal-content"><h5 class="{context_global.theme["modal_text"]}">{message}</h5></div>'
-          f'<form method="POST"><input type="hidden" name="tab" value="{tab}">'
-            '<div class="modal-footer">'
-              f'<button name="{btn_name}" value="{btn_value}" class="btn waves-effect waves-light">YES</button>'
-              '<a class="modal-close waves-effect waves-green btn-flat">Cancel</a>'
-            '</div>'
-          '</form>'
-        '</div>'
-    )
-
-    return button
-
-@app.template_global()
-def create_decora_switch(name: str, value: str, enabled: int):
-
-    off = ' active' if not enabled else ''
-    on  = ' active' if enabled else ''
-
-    switch = (
-        f'<div class="col s3"><div class="row row-thin"><p class="multi-switch-label center">{value}</p></div>'
-        '<div class="row row-thin"><div class="multi-switch-container decora-switch">'
-        '<ul class="multi-switch">'
-            f'<li class="multi-switch-off{off}"><button name="{name}" value="{value}" onclick="updateCategory(this, 0)">'
-                '<i class="material-icons small">radio_button_unchecked</i></button></li>'
-            f'<li class="multi-switch-on{on}"><button name="{name}" value="{value}" onclick="updateCategory(this, 1)">'
-                '<i class="material-icons small">block</i></button></li>'
-        '</ul></div></div></div>'
-    )
-
-    return switch
-
-@app.template_global()
-def merge_items(a1, a2):
-    '''accepts 2 arguments of item or list and merges them into one list. int can be replaced with any singular object.
-
-        valid combinations.
-            (int, list)
-            (int, int)
-            (list,list)
-            (list, int)
-    '''
-    new_list = []
-
-    for arg in [a1, a2]:
-
-        if not isinstance(arg, str) and hasattr(arg, '__iter__'):
-            new_list.extend(arg)
-
-        else:
-            new_list.append(arg)
-
-    return new_list
-
-@app.template_global()
-def is_list(li, /) -> bool:
-    return isinstance(li, list)
-
-
-# ====================================
+# =================================
 # JINJA2 API - CUSTOM TEMPLATES
-# ====================================
+# =================================
 app.jinja_env.filters['itoip'] = itoip
 
 # =================================
@@ -929,56 +787,3 @@ if (server_type == 'development'):
             response.headers.add('Cache-Control', 'no-store')
 
         return response
-
-
-# =================================
-# SECURE MESSENGER - TEMPORARY LOCATION
-# =================================
-# messenger will act like a single page application in that the uri will always be /messenger and chat will show if the
-# user has been authenticated, otherwise a login screen will be displayed.
-@app.route('/messenger', methods=['GET', 'POST'])
-def messenger_login():
-    page_settings = {
-        'navi': False, 'login_btn': False, 'idle_timeout': False, 'login_error': ''
-    }
-
-    if not (user := authenticated_session()):
-
-        if (request.method == 'POST'):
-
-            authenticated, username, user_role = Authentication.user_login(specify_role='messenger')
-            if (authenticated):
-                return redirect(url_for('messenger_login'))
-
-            page_settings['login_error'] = 'Invalid Credentials. Please try again.'
-
-        return render_template('messenger/login.html', theme=context_global.theme, **page_settings)
-
-    # AUTHENTICATED USERS
-    return messenger_chat()
-
-    ## CALL MESSENGER ENTRYPOINT, PASS IN USER INFO TO LOAD MESSAGES, AND RENDER TEMPLATE.
-
-@user_restrict('messenger', 'admin', login_page='messenger')  # NOTE: admin is for testing purposes only
-def messenger_chat(session_info: dict) -> str:
-
-    active_user = session_info['user']
-
-    if (request.method == 'POST'):
-
-        if ('change_recipients' not in request.form):
-
-            if not messenger.send_message(active_user, request.form):
-                return 'fuck'
-
-    recipient, messages = messenger.get_messages(active_user, request.form)
-
-    # loading user chats, then rendering template.
-    page_settings = {
-        'session_info': session_info,
-        'contacts': messenger.get_user_list(active_user),
-        'to_user': recipient,
-        'messages': messages
-    }
-
-    return render_template('messenger/chat.html', theme=context_global.theme, **page_settings)
